# openapi.yaml
openapi: 3.0.3
info:
  title: HackNext REST API
  description: Offline-first production hackathon platform API for collegiate engineering sprints.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local self-hosted backend server
paths:
...
  /events:
    get:
      summary: List all published hackathons
      responses:
        '200':
          description: Array of events
    post:
      summary: Create new hackathon event (Initializes 5-step timeline and default 4 rubrics)
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Created event object
...
  /rubrics/events/{eventId}:
    get:
      summary: Get evaluation criteria rubrics for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Array of criteria rubrics (sum of weights = 100%)
...
  /certificates/events/{eventId}/generate:
    post:
      summary: Generate SVG certificates locally on disk with SHA-256 integrity hash
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Generation completed
  /certificates/verify/{certificateNo}:
    get:
      summary: Verify certificate authenticity and integrity hash offline
      parameters:
        - name: certificateNo
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Certificate verification details and vector SVG preview
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT